pyproject.toml
[tool.poetry]
name = "skate_tricks"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
SQLAlchemy = "^1.4.31"
pydantic = "^1.9.0"
psycopg2-binary = "^2.9.3"
fastapi = "^0.73.0"
uvicorn = "^0.17.4"
SQLAlchemy-Utils = "^0.38.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pre-commit = "^2.17.0"
black = "^22.1.0"
mypy = "^0.931"
"testing.postgresql" = "^1.3.0"
alchemy-mock = "^0.4.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
